+++ src/components/SearchBar.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const SearchBar = () => {
+  return (
+    <div>SearchBar</div>
+  )
+}
+
+export default SearchBar
