+++ src/hooks/useFetch.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const useFetch = () => {
+
+}
+
+export default useFetch
