+++ src/pages/Blog.jsx
@@ -1,8 +1,15 @@
 import React from 'react'
+import { useParams } from 'react-router'
+import Header from '../components/Header'
 
 const Blog = () => {
-  return (
-    <div>Blog</div>
+  const {id} = useParams()
+
+  return (   
+    <div>
+      <Header />
+      <h1>This is Blog{id}</h1>
+    </div>
   )
 }
